Add new component: Enhance Positional Audio (#24)
While the EEs do not support EAX, positional audio still works through OpenAL, the default API. Unfortunately, due to the lack of configurability, many find the listening experience inadequate, especially with headphones. This is mostly because of the exaggerated panning that splits sounds into either the left or right channel.

This component adds the configurable software implementation of the API, OpenAL Soft, with crossfeed filters enabled to mitigate the channel separation.

Intended for BG:EE, SoD, BG2:EE, and EET.